9 move file redis_repository, abc_redis_repository to folder redis_cache
directory redis_cache respect to logic of redis' cached data
