+++ app_knife/repositories/knife_repository.py
@@ -1,12 +1,12 @@
 from sqlalchemy.orm import Session
 from redis import Redis
-from abc_handle_redis import AbcHandleRedis
-from handle_redis import HandleRedis
+from abstracts.abc_handle_redis import AbcHandleRedis
+from handlers.handle_redis import HandleRedis
 from app_knife.abstracts.abc_product_repository import AbcProductRepository
-from app_knife.abstracts.abc_redis_repository import AbcRedisRepository
+from abstracts.abc_redis_repository import AbcRedisRepository
 from app_knife.abstracts.abc_knife_repository import AbcKnifeRepository
 from app_knife.repositories.product_repository import ProductRepository
-from app_knife.repositories.redis_repository import RedisRepository
+from repositories.redis_repository import RedisRepository
 
 class KnifeRepository(AbcKnifeRepository):
